+++ RELEASENOTES.md
@@ -1,5 +1,9 @@
 # grunt-blue-button.js Release Notes
 
+# v.1.4.0
+- blue-button-generate task added.
+- blue-button-jsondiff task added.
+
 # v.1.3.0 - December 14, 2014
 - Creates Blue Button JSON files from XML files.
 - Optionally validates.